Replaces capistrano role with basic tasks
We were using a role from ansible-galaxy that used a module (or variable) that
was incompatible with Ansible v2. We replaced the role with simple tasks, since
all we need to do is create a directory anyway.

Still, this was tested against Ansbile v2 and it works.

Also...
* Specifies 'smart' gathering of facts, which reduces time to gather facts,
particularly nice when running single plays or tasks using --tags.
* Adds site.retry to gitignore. I think this only shows up when a playbook
fails.